@@ -4,16 +4,18 @@
   "main": "index.js",
   "license": "MIT",
   "scripts": {
-    "start": "webpack-dev-server --mode development"
+    "start": "webpack-dev-server --mode development",
+    "serve": "webpack serve",
+    "build": "webpack"
   },
   "dependencies": {
-    "rxjs": "^5.5.7",
-    "ts-loader": "^4.1.0",
-    "typescript": "^2.7.2",
-    "webpack": "^4.2.0",
-    "webpack-dev-server": "^3.1.1"
+    "rxjs": "^6.6.3",
+    "ts-loader": "^8.0.6",
+    "typescript": "^4.0.3",
+    "webpack": "^5.1.0",
+    "webpack-dev-server": "^3.11.0"
   },
   "devDependencies": {
-    "webpack-cli": "^2.0.13"
+    "webpack-cli": "^4.0.0"
   }
 }
