@@ -1,3 +1,4 @@
-import * as Rx from "rxjs/Observable";
+import * as Rx from "rxjs";
 
 console.log(Rx);
+console.log("Good");
